fix: Change COUNT(*) to return Integer instead of Numeric
- Modified COUNT(*) to return SqlValue::Integer in three key locations:
  - Fast path for simple COUNT(*) queries
  - COUNT(*) within aggregation context
  - AggregateAccumulator::finalize() method

- Updated all affected tests to expect Integer from COUNT(*)

- Note: Integer arithmetic remains Integer (no automatic promotion to Numeric)

Fixes #1199

+++ crates/executor/src/select/executor/aggregation/mod.rs
@@ -33,7 +33,7 @@ impl SelectExecutor<'_> {
             // If table doesn't exist, fall through to normal path which will produce proper error
             if let Some(table) = self.database.get_table(&table_name) {
                 let count = table.row_count();
-                return Ok(vec![storage::Row::new(vec![types::SqlValue::Numeric(count as f64)])]);
+                return Ok(vec![storage::Row::new(vec![types::SqlValue::Integer(count as i64)])]);
             }
         }
